change xstrsize_t from uint8_t to uint16_t, 'cause with uint8_t we were wasting space because of the padding

+++ internal/core/eonos/inc/xstr.h
@@ -48,7 +48,7 @@
 // ===============================================================================
 
 // type of xstr size
-typedef uint8_t xstrsize_t;
+typedef uint16_t xstrsize_t;
 
 // xstr main type
 typedef struct {
